made changes to reviews

+++ Taskr.Handlers/Task/DeleteJobHandler.cs
@@ -1,9 +1,11 @@
-﻿using System.Net;
+﻿using System.Linq;
+using System.Net;
 using System.Threading;
 using System.Threading.Tasks;
 using MediatR;
 using Microsoft.EntityFrameworkCore;
 using Taskr.Commands.Task;
+using Taskr.Domain;
 using Taskr.Dtos.Errors;
 using Taskr.Infrastructure.Security;
 using Taskr.Persistance;
@@ -25,7 +27,7 @@ namespace Taskr.Handlers.Task
         public async Task<Unit> Handle(DeleteJobCommand request, CancellationToken cancellationToken)
         {
             var loggedInUserId = _userAccess.GetCurrentUserId();
-            var job = await _context.Jobs.SingleOrDefaultAsync(x => x.Id == request.JobId, cancellationToken: cancellationToken);
+            var job = await _context.Jobs.Include(x => x.Photos).Include(x => x.Watching).SingleOrDefaultAsync(x => x.Id == request.JobId, cancellationToken: cancellationToken);
             if (job == null)
             {
                 throw new RestException(HttpStatusCode.NotFound, new {error = "Job not found"});
@@ -34,8 +36,27 @@ namespace Taskr.Handlers.Task
             if (job.UserId != loggedInUserId)
                 throw new RestException(HttpStatusCode.Unauthorized,
                     new {error = "You are unauthorized to delete this content"});
+
+            if (job.JobStatus != JobStatus.Active)
+                throw new RestException(HttpStatusCode.Forbidden, new {error = "This task is currently in use"});
+            
+            // get all chats associated with the job
+            var jobChats = await _context.Chats.Include(x => x.Messages).Where(x => x.Job == job).ToListAsync(cancellationToken);
+            
+            // delete all orders
+            
+            _context.Orders.RemoveRange(await _context.Orders.Where(x => x.Job == job).ToListAsync(cancellationToken));
+
+            _context.Chats.RemoveRange(jobChats);
+            
+            // foreach (var chat in jobChats)
+            // { 
+            //     _context.Messages.RemoveRange(chat.Messages);
+            // }
             
             _context.Jobs.Remove(job);
+            
+            
             
             var deleted = await _context.SaveChangesAsync(cancellationToken) > 0;
             if (!deleted)
